+++ jest.config.js
@@ -1,9 +1,17 @@
-const path = require("path");
+const path = require('path')
 
 module.exports = {
   verbose: true,
   transform: {
-    "^.+\\.tsx?$": "ts-jest",
+    '^.+\\.tsx?$': 'ts-jest',
   },
-  setupFilesAfterEnv: [path.join(__dirname, "./testing/setup-tests.js")],
-};
+  setupFiles: [path.join(__dirname, './testing/setup-tests.js')],
+  coverageThreshold: {
+    global: {
+      branches: 100,
+      functions: 100,
+      lines: 100,
+      statements: 100,
+    },
+  },
+}