+++ plugins/export-worbook/src/plugin.spec.ts
@@ -1,5 +1,5 @@
-describe("exportWorkbook", () => {
-  it("empty test", () => {
-    expect(true).toBe(true);
-  });
-});
+describe('exportWorkbook', () => {
+  it('empty test', () => {
+    expect(true).toBe(true)
+  })
+})