+++ plugins/json-extractor/src/index.ts
@@ -1,2 +1,11 @@
-export * from './json.extractor'
-export * from './json.extractor.plugin'
+import { parseBuffer } from './parser'
+import { Extractor } from '@flatfile/util-extractor'
+
+export const JSONExtractor = () => {
+  return Extractor('.json', parseBuffer)
+}
+
+/*
+ * @deprecated use `JSONExtractor` instead
+ */
+export const jsonExtractorPlugin = JSONExtractor