+++ plugins/record-hook/src/index.ts
@@ -1,4 +1,4 @@
-export * from "./RecordHook";
-export * from "./record.hook.plugin";
+export * from './RecordHook'
+export * from './record.hook.plugin'
 
-export { FlatfileRecord } from "@flatfile/hooks";
+export { FlatfileRecord } from '@flatfile/hooks'