+++ plugins/psv-extractor/CHANGELOG.md
@@ -1,5 +1,57 @@
 # @flatfile/plugin-psv-extractor
 
+## 1.3.3
+
+### Patch Changes
+
+- 2e7fce8: Remove unsupported parameter
+
+## 1.3.2
+
+### Patch Changes
+
+- Updated dependencies [1cd2f31]
+  - @flatfile/plugin-delimiter-extractor@0.3.0
+
+## 1.3.1
+
+### Patch Changes
+
+- 4651120: Add @deprecated comment
+
+## 1.3.0
+
+### Minor Changes
+
+- 77b9237: Add backwards compatibility for renamed extractors
+
+### Patch Changes
+
+- Updated dependencies [77b9237]
+  - @flatfile/plugin-delimiter-extractor@0.2.0
+
+## 1.2.0
+
+### Minor Changes
+
+- 2b15b32: Refactor extractors to share a common extractor and use custom parsers. Add .tsv file extractor and a generic delimiter extractor. Add header row detection to the Excel extractor.
+
+### Patch Changes
+
+- Updated dependencies [2b15b32]
+  - @flatfile/plugin-delimiter-extractor@0.1.0
+
+## 1.1.0
+
+### Minor Changes
+
+- e7e5134: DRYing abstract extractor code
+
+### Patch Changes
+
+- Updated dependencies [e7e5134]
+  - @flatfile/plugin-extractor-utils@0.1.0
+
 ## 1.0.0
 
 ### Major Changes