+++ plugins/xlsx-extractor/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @flatfile/plugin-record-hook
 
+## 1.4.0
+
+### Minor Changes
+
+- 6189d31: Fix for ghost rows in Excel files
+
+## 1.3.2
+
+### Patch Changes
+
+- 1cd2f31: Add header row auto-detection
+- Updated dependencies [c859a10]
+  - @flatfile/util-extractor@0.2.1
+
+## 1.3.1
+
+### Patch Changes
+
+- 4651120: Add @deprecated comment
+
+## 1.3.0
+
+### Minor Changes
+
+- 77b9237: Add backwards compatibility for renamed extractors
+
+## 1.2.0
+
+### Minor Changes
+
+- 2b15b32: Refactor extractors to share a common extractor and use custom parsers. Add .tsv file extractor and a generic delimiter extractor. Add header row detection to the Excel extractor.
+
+### Patch Changes
+
+- Updated dependencies [2b15b32]
+  - @flatfile/util-extractor@0.2.0
+
+## 1.1.0
+
+### Minor Changes
+
+- e7e5134: DRYing abstract extractor code
+
+### Patch Changes
+
+- Updated dependencies [e7e5134]
+  - @flatfile/plugin-extractor-utils@0.1.0
+
 ## 1.0.8
 
 ### Patch Changes